feat: add remove item feature both in grid and list view layout
