--- lib/file/presentation/view/widgets/file_index_body.dart
@@ -1,408 +1,0 @@
-// ignore_for_file: prefer_const_constructors
-
-import 'dart:convert';
-import 'dart:math';
-import 'dart:typed_data';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:get_it/get_it.dart';
-import 'package:ui_library/ui_library_export.dart';
-import 'package:uplink/file/domain/item.dart';
-import 'package:uplink/file/presentation/controller/item_list_bloc.dart';
-import 'package:uplink/file/presentation/view/helper/show_file_options_bottom_sheet.dart';
-import 'package:uplink/file/presentation/view/remove_list_page/models/selected_item_list.dart';
-import 'package:uplink/file/presentation/view/widgets/widgets_export.dart';
-import 'package:uplink/l10n/main_app_strings.dart';
-
-class FileIndexBody extends StatefulWidget {
-  const FileIndexBody.itemList({
-    super.key,
-  })  : isFavoritesList = false,
-        isReomveList = false;
-
-  const FileIndexBody.favoritesList({
-    super.key,
-  })  : isFavoritesList = true,
-        isReomveList = false;
-
-  const FileIndexBody.removeList({
-    super.key,
-  })  : isFavoritesList = false,
-        isReomveList = true;
-
-  final bool isFavoritesList;
-  final bool isReomveList;
-
-  @override
-  State<FileIndexBody> createState() => _FileIndexBodyState();
-}
-
-class _FileIndexBodyState extends State<FileIndexBody> {
-  final _itemListController = GetIt.I.get<ItemListBloc>();
-  late bool gridViewIsSelected;
-  @override
-  void initState() {
-    super.initState();
-    gridViewIsSelected = true;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-//Item layout bar
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 0, 10, 0),
-          child: SizedBox(
-            height: 40,
-            child: Row(
-              children: [
-                Expanded(
-                  child: BlocBuilder<ItemListBloc, ItemListState>(
-                    bloc: _itemListController,
-                    builder: (context, state) {
-                      if (state is ItemListLoadSuccess) {
-                        final int _listLength;
-
-                        if (widget.isFavoritesList == true) {
-                          final _favoritesList = state.itemList
-                              .where((element) => element.isFavorited == true);
-                          _listLength = _favoritesList.length;
-                        } else {
-                          _listLength = state.itemList.length;
-                        }
-// # of items
-                        return UText(
-                          state.itemList.isEmpty
-                              ? UAppStrings.file_index_page_no_items
-                              : _listLength.toString() +
-                                  UAppStrings.file_index_page_items,
-                          textStyle: UTextStyle.H5_fifthHeader,
-                        );
-                      } else if (state is ItemListLoadFailure) {
-                        return const UText(
-                          UAppStrings.file_index_page_items_in_error,
-                          textStyle: UTextStyle.H5_fifthHeader,
-                        );
-                      }
-                      return const UShimmer.med(
-                        child: URectContainer(
-                          height: 18,
-                        ),
-                      );
-                    },
-                  ),
-                ),
-                const SizedBox(width: 16),
-                ItemLayoutButton(
-                  onPressed: () {
-                    setState(() {
-                      gridViewIsSelected = true;
-                    });
-                  },
-                  isSelected: gridViewIsSelected,
-                  uIconData: UIcons.grid_view,
-                ),
-                const SizedBox(width: 12),
-                ItemLayoutButton(
-                  onPressed: () {
-                    setState(() {
-                      gridViewIsSelected = false;
-                    });
-                  },
-                  isSelected: !gridViewIsSelected,
-                  uIconData: UIcons.list_view,
-                ),
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(height: 4),
-// photos list
-        Expanded(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16),
-            child: BlocBuilder<ItemListBloc, ItemListState>(
-              bloc: _itemListController,
-              builder: (context, state) {
-                if (state is ItemListLoadSuccess) {
-                  // favorites list
-                  if (widget.isFavoritesList == true) {
-                    final _favoritesList = state.itemList
-                        .where((element) => element.isFavorited == true)
-                        .toList();
-                    if (_favoritesList.isEmpty) {
-                      return Text("You don't have any favorited item");
-                    }
-
-                    return gridViewIsSelected
-                        ? ItemGridView(
-                            itemList: _favoritesList,
-                          )
-                        : ItemListView(
-                            itemList: _favoritesList,
-                          );
-                  } else if (widget.isReomveList == true) {
-                    // remove list
-                    final _removeList = state.itemList;
-
-                    if (_removeList.isEmpty) {
-                      return Text("You don't have any item to remove");
-                    }
-
-                    return gridViewIsSelected
-                        ? ItemGridView(
-                            itemList: _removeList,
-                            isRemoveList: true,
-                          )
-                        : ItemListView(
-                            itemList: _removeList,
-                            isRemoveList: true,
-                          );
-                  } else {
-                    // item list
-                    if (state.itemList.isEmpty) {
-                      return const NoItemBody();
-                    }
-                    return gridViewIsSelected
-                        ? ItemGridView(
-                            itemList: state.itemList,
-                          )
-                        : ItemListView(
-                            itemList: state.itemList,
-                          );
-                  }
-                } else if (state is ItemListLoadFailure) {
-                  return const UText(
-                    UAppStrings.file_index_page_load_in_error,
-                    textStyle: UTextStyle.H5_fifthHeader,
-                  );
-                }
-                // Initial and loading state
-                return const LoadingGridView();
-              },
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class ItemListView extends StatefulWidget {
-  const ItemListView({
-    Key? key,
-    required this.itemList,
-    this.isRemoveList = false,
-  }) : super(key: key);
-
-  final List<Item> itemList;
-  final bool isRemoveList;
-
-  @override
-  State<ItemListView> createState() => _ItemListViewState();
-}
-
-class _ItemListViewState extends State<ItemListView> {
-  @override
-  Widget build(BuildContext context) {
-    return ListView.separated(
-      itemCount: widget.itemList.length,
-      itemBuilder: (context, index) {
-        final _item = widget.itemList[index];
-        Uint8List? _itemUnit8ListImage;
-        if (widget.itemList[index].thumbnail != null) {
-          _itemUnit8ListImage = base64Decode(widget.itemList[index].thumbnail!);
-        }
-
-        return ListViewListTile(
-          item: _item,
-          itemUnit8ListImage: _itemUnit8ListImage,
-          isRemoveList: widget.isRemoveList,
-        );
-      },
-      separatorBuilder: (context, index) => const SizedBox(
-        height: 24,
-      ),
-    );
-  }
-}
-
-class ListViewListTile extends StatefulWidget {
-  const ListViewListTile({
-    Key? key,
-    required Item item,
-    required Uint8List? itemUnit8ListImage,
-    required bool isRemoveList,
-  })  : _item = item,
-        _itemUnit8ListImage = itemUnit8ListImage,
-        _isRemoveList = isRemoveList,
-        super(key: key);
-
-  final Item _item;
-  final Uint8List? _itemUnit8ListImage;
-  final bool _isRemoveList;
-
-  @override
-  State<ListViewListTile> createState() => _ListViewListTileState();
-}
-
-class _ListViewListTileState extends State<ListViewListTile> {
-  bool _isSelected = false;
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: UColors.backgroundDark,
-      shape: RoundedRectangleBorder(
-        side: const BorderSide(color: UColors.foregroundDark),
-        borderRadius: BorderRadius.circular(4),
-      ),
-      margin: EdgeInsets.zero,
-      child: ListTile(
-        dense: true,
-        visualDensity: const VisualDensity(vertical: -1),
-        leading: widget._item.type == ItemType.photo
-            ? Container(
-                height: 40,
-                width: 40,
-                decoration: BoxDecoration(
-                  borderRadius: const BorderRadius.all(
-                    Radius.circular(4),
-                  ),
-                  image: DecorationImage(
-                    image: MemoryImage(widget._itemUnit8ListImage!),
-                    fit: BoxFit.cover,
-                  ),
-                ),
-              )
-            // TODO(yijing): add thumbnail for non-photo item
-            : Container(
-                height: 40,
-                width: 40,
-                color: UColors.foregroundDark,
-              ),
-        title: UText(
-          widget._item.name,
-          textStyle: UTextStyle.H5_fifthHeader,
-          textColor: UColors.white,
-        ),
-        // TODO(yijing): update size unit
-        subtitle: UText(
-          widget._item.size.toFileSize(decimals: 1),
-          textStyle: UTextStyle.M1_micro,
-        ),
-        trailing: widget._isRemoveList
-            ? IconButton(
-                icon: UIcon(
-                  _isSelected == false ? UIcons.select_box : UIcons.checkmark_1,
-                ),
-                onPressed: () {
-                  if (widget._isRemoveList == true) {
-                    if (_isSelected == false) {
-                      setState(() {
-                        _isSelected = true;
-                        // context.read<SelectedItemList>().add(_item);
-                      });
-                    } else {
-                      setState(() {
-                        _isSelected = false;
-                        // context.read<SelectedItemList>().remove(_item);
-                      });
-                    }
-                  } else {
-                    showFileOptionsBottomSheet(context, widget._item);
-                  }
-                },
-                visualDensity: const VisualDensity(horizontal: -2),
-              )
-            : IconButton(
-                icon: UIcon(
-                  UIcons.hamburger_menu,
-                ),
-                onPressed: () {
-                  showFileOptionsBottomSheet(context, widget._item);
-                },
-                visualDensity: const VisualDensity(horizontal: -2),
-              ),
-        contentPadding: const EdgeInsets.fromLTRB(8, 0, 0, 0),
-      ),
-    );
-  }
-}
-
-class ItemGridView extends StatefulWidget {
-  const ItemGridView({
-    Key? key,
-    required this.itemList,
-    this.isRemoveList = false,
-  }) : super(key: key);
-
-  final List<Item> itemList;
-  final bool isRemoveList;
-
-  @override
-  State<ItemGridView> createState() => ItemGridViewState();
-}
-
-class ItemGridViewState extends State<ItemGridView> {
-  @override
-  Widget build(BuildContext context) {
-    return GridView.builder(
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: 2,
-        mainAxisSpacing: 8,
-        crossAxisSpacing: 8,
-        childAspectRatio: 160 / 168,
-      ),
-      itemCount: widget.itemList.length,
-      itemBuilder: (context, index) {
-        final _item = widget.itemList[index];
-        final _imageUint8List = base64.decode(_item.thumbnail!);
-        if (widget.itemList[index].type == ItemType.photo) {
-          final _item = widget.itemList[index];
-
-          // Item in the RemoveListPage
-          if (widget.isRemoveList == true) {
-            return UImageButton.unit8ListImage(
-              unit8ListImage: _imageUint8List,
-              isFavored: _item.isFavorited,
-              isDeleting: true,
-              onSelected: () {
-                context.read<SelectedItemList>().add(_item);
-              },
-              unSelected: () {
-                context.read<SelectedItemList>().remove(_item);
-              },
-            );
-          }
-
-          // Item in the FileIndexPage and FavoriteListPage
-          return GestureDetector(
-            onTap: () {
-              Navigator.of(context)
-                  .pushNamed('/PhotoDetailPage', arguments: _item);
-            },
-            onLongPress: () {
-              showFileOptionsBottomSheet(context, _item);
-            },
-            child: UImageButton.unit8ListImage(
-              unit8ListImage: _imageUint8List,
-              isFavored: _item.isFavorited,
-            ),
-          );
-        }
-        // TODO(yijing): show thumbnail when item is not photo
-        return const SizedBox.shrink();
-      },
-    );
-  }
-}
-
-extension on int {
-  String toFileSize({int decimals = 0}) {
-    final suffixes = ['B', 'KB', 'MB', 'GB', 'TB'];
-    final i = (log(this) / log(1024)).floor();
-    return '${(this / pow(1024, i)).toStringAsFixed(decimals)} ${suffixes[i]}';
-  }
-}
