+++ lib/file/presentation/view/widgets/widgets_export.dart
@@ -1,5 +1,5 @@
 export 'file_index_app_bar.dart';
-export 'file_index_body.dart';
+export 'file_index_body/file_index_body.dart';
 export 'item_layout_button.dart';
 export 'loading_gridview.dart';
 export 'no_item_body.dart';
